padre/cdi/cdi-ejemplos/src/test/java/pe/joedayz/javaee7/cdi/ejemplo19/DerbyPingService19Test.java
package pe.joedayz.javaee7.cdi.ejemplo19;

import org.jboss.weld.environment.se.Weld;
import org.jboss.weld.environment.se.WeldContainer;
import org.junit.AfterClass;
import org.junit.BeforeClass;
import org.junit.Test;

import java.sql.Connection;
import java.sql.DriverManager;

import static org.junit.Assert.assertTrue;


public class DerbyPingService19Test {

  // ======================================
  // =             Attributes             =
  // ======================================

  protected static Weld weld;
  protected static WeldContainer container;

  // ======================================
  // =          Lifecycle Methods         =
  // ======================================

  @BeforeClass
  public static void init() {
    weld = new Weld();
    container = weld.initialize();
  }

  @AfterClass
  public static void close() {
    weld.shutdown();
  }

  // ======================================
  // =              Methods               =
  // ======================================

  @Test
  public void shouldPingDatabase() throws Exception {
    Class.forName("org.apache.derby.jdbc.EmbeddedDriver").newInstance();
    Connection conn = DriverManager.getConnection("jdbc:derby:memory:cdiDB;create=true", "APP", "APP");
    conn.createStatement().executeQuery("SELECT 1 FROM SYSIBM.SYSDUMMY1");
    conn.close();
  }

  @Test
  public void shouldPingDatabaseWithDispose() throws Exception {
    DerbyPingService19 pingService = container.instance().select(DerbyPingService19.class).get();
    pingService.ping();
  }
}